Moved device farm functions into separate package and created sylph api.
